feat(OrderManagement): Enhance order management UI and functionality
- Update route names for better clarity in Vietnamese, including adjustments to order details and invoice labels.
- Refactor User interface to utilize enums for role and status, improving type safety and maintainability.
- Introduce new components for order filtering and searching, enhancing user experience in managing orders.
- Implement loading and empty states for order views, improving feedback during data fetching.
- Create OrderCardSkeleton for loading states and integrate it into the order management flow.
